feature/DEV_ENV-improvement of development environment -- simplify example
