+++ app-example-js/src/index.js
@@ -1,13 +1,38 @@
 
 // outsource dependencies
-import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
+import React, { useCallback, useEffect } from 'react';
 
 // local dependencies
 import './index.css';
-import App from './App';
 import store from './store';
+import logo from './logo.svg';
+import { controller } from './controller';
+import { useController } from 'redux-saga-controller';
+
+
+function App() {
+  const [
+    { data, disabled, initialized },
+    { initialize, getSelf },
+  ] = useController(controller);
+
+  useEffect(() => { initialize({ some: 'probably route data for initialization'}) }, [initialize]);
+  const handleGetData = useCallback(getSelf, [getSelf]);
+
+  return <div className="container">
+    { !initialized ? <img src={logo} className="logo" alt="logo" /> : <>
+      <h2> DATA </h2>
+      <p> { JSON.stringify(data) } </p>
+      <button className="btn" disabled={disabled} onClick={handleGetData}>
+        { disabled ? <img src={logo} className="logo" alt="logo" /> : <span> GET </span> }
+        {/*<img src={logo} className="logo" alt="logo" />*/}
+      </button>
+    </> }
+  </div>;
+}
+
 
 ReactDOM.render(
   <React.StrictMode>
